v_7.3.4

+++ src/main/resources/db/migration/V05__alterar_prontuario_criar_consulta.sql
@@ -1,7 +1,7 @@
 create table consulta (
 	id serial primary key,
 	codigo_prontuario integer not null,
-	cpf_medico integer not null,
+	cpf_medico BIGINT not null,
 	prescricao TEXT
 );
